Added a bash check to see which system I am on so that I no longer worry about using different OSs with this bashrc.
Added a check to see if git-prompt.sh is there to source.

+++ bashrc
@@ -1,22 +1,42 @@
-if [ -f `brew --prefix`/etc/bash_completion ]; then
-    . `brew --prefix`/etc/bash_completion
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    # ...
+    alias ls='ls --color=auto'
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    # Mac OSX
+    if [ -f `brew --prefix`/etc/bash_completion ]; then
+        . `brew --prefix`/etc/bash_completion
+    fi
+    alias ls='ls -G'
+    alias flush='dscacheutil -flushcache && sudo arp -d -a && sudo killall -HUP mDNSResponder'
+elif [[ "$OSTYPE" == "cygwin" ]]; then
+    # POSIX compatibility layer and Linux environment emulation for Windows
+    echo $OSTYPE
+elif [[ "$OSTYPE" == "msys" ]]; then
+    # Lightweight shell and GNU utilities compiled for Windows (part of MinGW)
+    echo $OSTYPE
+elif [[ "$OSTYPE" == "win32" ]]; then
+    # I'm not sure this can happen.
+    echo $OSTYPE
+elif [[ "$OSTYPE" == "freebsd"* ]]; then
+    # ...
+    echo $OSTYPE
+else
+    # Unknown.
+    echo $OSTYPE
 fi
 
 export CLICOLOR=1
 export LSCOLORS=gxBxhxDxfxhxhxhxhxcxcx
 
-source ~/.git-prompt.sh
-
+[[ -s "$HOME/.git-prompt.sh" ]] && source $HOME/.git-prompt.sh
 [[ -s "$HOME/.profile" ]] && source "$HOME/.profile" # Load the default .profile
 [[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm" # Load RVM into a shell session *as a function*
 
-alias ls='ls -a'
-alias vi='vim'
-alias flush='dscacheutil -flushcache && sudo arp -d -a && sudo killall -HUP mDNSResponder'
- 
+#alias vi='vim'
+
 # Control History
 export EDITOR=vim
-export ARCHFLAGS="-arch x86_64"
+#export ARCHFLAGS="-arch x86_64"
 
 # Reset
 Color_Off="\[\033[0m\]"       # Text Reset
@@ -101,16 +121,16 @@ Jobs="\j"
 
 
 export PS1=$IBlue$Time12h$Color_Off'$(git branch &>/dev/null;\
-if [ $? -eq 0 ]; then \
-  echo "$(echo `git status` | grep "nothing to commit" > /dev/null 2>&1; \
-  if [ "$?" -eq "0" ]; then \
-    # @4 - Clean repository - nothing to commit
-    echo "'$Green'"$(__git_ps1 " (%s)"); \
-  else \
-    # @5 - Changes to working tree
-    echo "'$IRed'"$(__git_ps1 " {%s}"); \
-  fi) '$BYellow$PathShort$Color_Off'\$ "; \
-else \
-  # @2 - Prompt when not in GIT repo
-  echo " '$Yellow$PathShort$Color_Off'\$ "; \
+    if [ $? -eq 0 ]; then \
+        echo "$(echo `git status` | grep "nothing to commit" > /dev/null 2>&1; \
+        if [ "$?" -eq "0" ]; then \
+            # @4 - Clean repository - nothing to commit
+            echo "'$Green'"$(__git_ps1 " (%s)"); \
+        else \
+            # @5 - Changes to working tree
+            echo "'$IRed'"$(__git_ps1 " {%s}"); \
+        fi) '$BYellow$PathShort$Color_Off'\$ "; \
+    else \
+        # @2 - Prompt when not in GIT repo
+        echo " '$Yellow$PathShort$Color_Off'\$ "; \
 fi)'
